+++ .gen/public/model/user_orders.go
@@ -12,7 +12,7 @@ import (
 )
 
 type UserOrders struct {
-	ID        int32 `sql:"primary_key"`
+	ID        int64 `sql:"primary_key"`
 	UserID    *int64
 	ItemID    *int64
 	Status    *string
